project/source/requirements.txt
Django==1.11.16
Faker==0.9.2
Pillow==5.3.0
PyJWT==1.6.4
argparse==1.2.1
boto3==1.9.27
botocore==1.12.27
certifi==2018.10.15
chardet==3.0.4
concurrent-log-handler==0.9.12
confusable-homoglyphs==3.2.0
debug-tools==2.2.1
django-crispy-forms==1.7.2
django-crispy-forms-registration==0.1.3
django-debug-toolbar==1.10.1
django-factory-boy==1.0.0
django-json-rpc==0.7.2
django-registration==3.0
django-storages==1.7.1
docutils==0.14
factory-boy==2.11.1
funcsigs==1.0.2
futures==3.2.0
idna==2.7
ipaddress==1.0.22
jmespath==0.9.3
mock==2.0.0
oauthlib==2.1.0
pbr==4.3.0
pip==18.0
ez-setup==0.9
portalocker==1.2.1
python-dateutil==2.7.3
python-magic==0.4.15
python-openid==2.2.5
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
s3transfer==0.1.13
setuptools==40.4.3
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sqlparse==0.2.4
text-unidecode==1.2
mysqlclient==1.3.13
urllib3==1.23
wheel==0.32.0
wsgiref==0.1.2










